requirements.txt
cat > requirements.txt << EOF
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
numpy==1.24.3
scikit-learn==1.3.2
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.35.0
groq==0.4.1
gunicorn==21.2.0
python-dotenv==1.0.0
EOF